DCF_Backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
daphne==4.1.2
Deprecated==1.2.15
distlib==0.3.8
distro==1.9.0
dj-database-url==2.3.0
Django==5.0.2
django-anymail==10.3
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-smtp-ssl==1.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
et-xmlfile==1.1.0
filelock==3.13.1
flatbuffers==24.12.23
fonttools==4.54.1
frozenlist==1.4.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
humanize==4.5.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
keyboard==0.13.5
kiwisolver==1.4.7
libclang==18.1.1
libmagic==1.0
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.9.4
numpy==2.0.0
openai==1.0.0
opencv-python==4.10.0.84
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pi==0.1.2
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
PyGithub==1.57
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.7
python-dotenv==1.0.1
python-gitlab==3.13.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
reviews==0.6.0
rich==13.3.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
statsmodels==0.14.4
streamlit==1.39.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
watchdog==5.0.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.15.5
zope.interface==7.1.0
